feat: 도메인 단위로 패키징 구성 수정
