+++ src/main/java/gift/product/application/ProductService.java
@@ -1,0 +1,67 @@
+package gift.product.application;
+
+import gift.product.domain.Product;
+import gift.product.exception.ProductException;
+import gift.product.infra.ProductRepository;
+import gift.product.presentation.WishListManageController.CreateProductRequestDTO;
+import gift.util.ErrorCode;
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ProductService {
+
+    @Autowired
+    private ProductRepository productRepository;
+    private static final int MAX_PRODUCT_NAME_LENGTH = 15;
+
+
+    public void addProduct(CreateProductRequestDTO createProductRequestDTO) {
+        Product product = new Product(createProductRequestDTO.name(), createProductRequestDTO.price(),
+            createProductRequestDTO.imageUrl());
+        validateProduct(product);
+        productRepository.addProduct(product);
+    }
+
+    public void deleteProduct(Long id) {
+        productRepository.deleteProduct(id);
+    }
+
+    public void updateProduct(Long id, String name, Double price, String imageUrl) {
+        Product product = new Product(name, price, imageUrl);
+        productRepository.updateProduct(id, product);
+    }
+
+    private void validateProduct(Product product) {
+        validateName(product.name());
+        validatePrice(product.price());
+    }
+
+    private void validateName(String name) {
+        if (name == null || name.isEmpty()) {
+            throw new ProductException(ErrorCode.INVALID_NAME);
+        }
+        if (name.length() > MAX_PRODUCT_NAME_LENGTH) {
+            throw new ProductException(ErrorCode.NAME_TOO_LONG);
+        }
+    }
+
+    private void validatePrice(Double price) {
+        if (price == null) {
+            throw new ProductException(ErrorCode.INVALID_PRICE);
+        }
+        if (price < 0) {
+            throw new ProductException(ErrorCode.NEGATIVE_PRICE);
+        }
+    }
+
+    public Product getProductByName(Long id) {
+        return productRepository.getProductById(id);
+    }
+
+    public List<Product> getProduct() {
+        return productRepository.getProducts();
+    }
+
+}
